Fix User-Agent
The first version of the PHP proxy actually didn't set the User-Agent
properly due to me forgetting a `User-Agent: ` in front of `Homebr3w/`.
This is fixed.

Unfortunately, there is still an issue in getcia.php that requires
attention. For some reason downloading a `.cia` with `file_get_contents`
works fine, but trying to use the `http_req` curl function only returns
`[]`.