Add apps update check
Based on mtime, which is the Unix Epoch Timestamp TitleDB.com has found
the newest version of an app at, Homebr3w now detects if a new update
for an app came out and will tell the user about it.
For those that have installed homebrew apps from other sources, Homebr3w
will tell them to reinstall the app so the mtime can be set
appropriately or the User has the chance to mark all unknown versions as
latest in the menu.

Colors:
BLUE: Only top screen, shows the currently selected item
GREEN: The latest version of the app is installed
RED: An outdated version of the app is installed
YELLOW: Version unknown, the app is installed, but no mtime been found
WHITE: The app is not installed