Better handling of external libraries
Having to include libraries in a project is always a very bad idea, but
I did it anyways. This commit stands to correct that by not including
and instead downloading the most recent libraries from the author's
pages, also keeping things like copyright and license of the files
intact. It'll also make it easier to add new libraries by simply
defining a new entry in the libraries table.

Closes #7