Add UUID to app
Generate a unique UUID for a Homebr3w Client that gets stored on SD.
The UUID is then appended to every `getJSON()` and `getFile()` as the
`homebr3wUUID` GET parameter. This will help @ksanislo and me to add a
way of collecting pseudo-anonymous data.

The sending of the UUID can be disabled in the settings area of the app.
The UUID can be regenerated by deleting `data.json` or the `client_uuid`
entry inside.